pyproject.toml
[project]
name = "robotics-ai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "logfire>=4.3.6",
    "numpy>=2.2.6",
    "pandas>=2.3.2",
    "pydantic-ai[examples]>=0.8.1",
    "pydantic-graph>=0.8.1",
    "python-dotenv>=1.1.1",
]

